--- packages/checkout/widgets-lib/src/context/analytics-provider/AnalyticsProvider.tsx
@@ -1,13 +1,0 @@
-import { ReactNode } from 'react';
-import { AnalyticsProvider } from '../AnalyticsProvider';
-
-interface AnalyticsProviderProps {
-  children: ReactNode;
-}
-export function SegmentAnalyticsProvider({ children }: AnalyticsProviderProps) {
-  return (
-    <AnalyticsProvider>
-      {children}
-    </AnalyticsProvider>
-  );
-}
